refactor
moved the 2fa reminders to `/components` folder
