+++ meshapp/src/components/Layout/index.tsx
@@ -2,7 +2,7 @@ import { Box } from "@mui/material";
 import { NavBar } from "./Header/NavBar";
 import { Outlet } from "react-router-dom";
 import Footer from "./Footer/Footer";
-import TwoFactorAuthReminders from "../../two-factor-auth/two-factor-reminder";
+import TwoFactorAuthReminders from "../two-factor-reminder/two-factor-reminder";
 import MessageBar from "../MessageBar";
 
 export const PageLayout = () => {
