+++ meshapp/src/components/two-factor-reminder/styling/two-factor-banner.css
@@ -1,29 +1,31 @@
 .banner {
   box-sizing: border-box;
   width: 100%;
-  height: 50px;
+  height: fit-content;
   display: flex;
   flex-wrap: nowrap;
   justify-content: left;
   align-items: center;
-  padding: 1em;
-  background-color: rgb(35, 204, 170);
+  padding: 4px 12px;
+
   color: #fff;
   column-gap: 10px;
   position: absolute;
   top: 0px;
   z-index: 999;
-  box-shadow: 0px 1.5px 5px black;
+  box-shadow: 0px 1px 3px black;
+
+  justify-content: space-between;
+}
+
+.bannerContent {
+  display: flex;
+  align-items: center;
+  column-gap: 10px;
 }
 
 .bannerLogo {
   display: flex;
-}
-
-.bannerCloseBtn {
-  display: flex;
-  position: absolute;
-  right: 30px;
 }
 
 .descText {
@@ -34,4 +36,9 @@
 
 .bannerBtn {
   display: flex;
+}
+
+.bannerCloseBtn {
+  display: flex;
+  cursor: pointer;
 }
