+++ meshapp/src/components/two-factor-reminder/styling/two-factor-popup.css
@@ -31,7 +31,7 @@
 
 @font-face {
   font-family: "cocogoose";
-  src: url("../../../public/fonts/cocogoose.ttf");
+  src: url("../../../../public/fonts/cocogoose.ttf");
 }
 
 .text,
